requirements.txt
pytubefix
pythondot-env
